refactor(#804): Remove DecisionNode interface
