+++ generator/src/test/java/com/scottlogic/deg/generator/decisiontree/treepartitioning/ConstraintToFieldMapperTests.java
@@ -23,7 +23,6 @@ import com.scottlogic.deg.common.profile.constraints.atomic.AtomicConstraint;
 import com.scottlogic.deg.generator.decisiontree.DecisionNode;
 import com.scottlogic.deg.generator.decisiontree.DecisionTree;
 import com.scottlogic.deg.generator.decisiontree.ConstraintNode;
-import com.scottlogic.deg.generator.decisiontree.TreeDecisionNode;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.DistributedSet;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.WeightedElement;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.FrequencyDistributedSet;
@@ -57,7 +56,7 @@ class ConstraintToFieldMapperTests {
         givenFields("B");
 
         final AtomicConstraint constraint = new IsInSetConstraint(new Field("B"), whitelistOf("test-value"));
-        final DecisionNode decision = new TreeDecisionNode(
+        final DecisionNode decision = new DecisionNode(
             new ConstraintNode(constraint));
 
         givenDecisions(decision);
@@ -89,22 +88,22 @@ class ConstraintToFieldMapperTests {
         final AtomicConstraint constraintE = new IsInSetConstraint(new Field("E"), whitelistOf("test-value"));
         final AtomicConstraint constraintF = new IsInSetConstraint(new Field("F"), whitelistOf("test-value"));
 
-        final DecisionNode decisionABC = new TreeDecisionNode(
+        final DecisionNode decisionABC = new DecisionNode(
             new ConstraintNode(
                 Collections.emptyList(),
                 Arrays.asList(
-                    new TreeDecisionNode(new ConstraintNode(constraintA)),
-                    new TreeDecisionNode(new ConstraintNode(constraintB)),
-                    new TreeDecisionNode(new ConstraintNode(constraintC))
+                    new DecisionNode(new ConstraintNode(constraintA)),
+                    new DecisionNode(new ConstraintNode(constraintB)),
+                    new DecisionNode(new ConstraintNode(constraintC))
                 )
             )
         );
 
-        final DecisionNode decisionDEF = new TreeDecisionNode(
+        final DecisionNode decisionDEF = new DecisionNode(
             new ConstraintNode(
                 Collections.emptyList(),
                 Collections.singletonList(
-                    new TreeDecisionNode(
+                    new DecisionNode(
                         new ConstraintNode(constraintD),
                         new ConstraintNode(constraintE),
                         new ConstraintNode(constraintF))
